Rename FixedLengthTupleStoreBackend
