+++ great_expectations/data_context/store/html_site_store.py
@@ -7,7 +7,7 @@ from great_expectations.data_context.types.resource_identifiers import (
     ValidationResultIdentifier,
     SiteSectionIdentifier,
 )
-from .fixed_length_tuple_store_backend import TupleStoreBackend
+from .tuple_store_backend import TupleStoreBackend
 from great_expectations.data_context.util import (
     load_class,
     instantiate_class_from_config,
