+++ great_expectations/data_context/store/validations_store.py
@@ -1,7 +1,7 @@
 from great_expectations.core import ExpectationSuiteValidationResultSchema
 from great_expectations.data_context.store.database_store_backend import DatabaseStoreBackend
 from great_expectations.data_context.store.store import Store
-from great_expectations.data_context.store.fixed_length_tuple_store_backend import TupleStoreBackend
+from great_expectations.data_context.store.tuple_store_backend import TupleStoreBackend
 from great_expectations.data_context.types.resource_identifiers import ValidationResultIdentifier
 from great_expectations.data_context.util import load_class
